UI 상태 증가 구현 & 랜덤생성  미구현

+++ Assets/Scripts/Bullet Controller.cs
@@ -3,7 +3,7 @@ using UnityEngine.Pool;
 public class BulletController : MonoBehaviour
 {
     public float bulletSpeed;
-    public float playerBulletDamage;
+    public int playerBulletDamage;
 
     private IObjectPool<BulletController> _ManagePool;
     void Update()
